Refactor context item management: move ContextItemMap to settings.ts and update imports across components
